+++ galasa-ecosystem-kubernetes-operator/deploy/role_binding.yaml
@@ -1,11 +1,11 @@
 kind: RoleBinding
 apiVersion: rbac.authorization.k8s.io/v1
 metadata:
-  name: galasa-ecosystem-operator
+  name: galasa-ecosystem-kubernetes-operator
 subjects:
 - kind: ServiceAccount
-  name: galasa-ecosystem-operator
+  name: galasa-ecosystem-kubernetes-operator
 roleRef:
   kind: Role
-  name: galasa-ecosystem-operator
+  name: galasa-ecosystem-kubernetes-operator
   apiGroup: rbac.authorization.k8s.io
