+++ galasa-ecosystem-kubernetes-operator/deploy/service_account.yaml
@@ -1,4 +1,4 @@
 apiVersion: v1
 kind: ServiceAccount
 metadata:
-  name: galasa-ecosystem-operator
+  name: galasa-ecosystem-kubernetes-operator
