+++ galasa-ecosystem-kubernetes-operator/pkg/apis/addtoscheme_galasa_v1alpha1.go
@@ -1,7 +1,7 @@
 package apis
 
 import (
-	"github.com/galasa-dev/extensions/galasa-ecosystem-operator/pkg/apis/galasa/v1alpha1"
+	"github.com/galasa-dev/extensions/galasa-ecosystem-kubernetes-operator/pkg/apis/galasa/v1alpha1"
 )
 
 func init() {
