+++ galasa-ecosystem-kubernetes-operator/pkg/cps/cps_test.go
@@ -3,7 +3,7 @@ package cps
 import (
 	"testing"
 
-	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-operator/pkg/apis/galasa/v1alpha1"
+	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-kubernetes-operator/pkg/apis/galasa/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
