+++ galasa-ecosystem-kubernetes-operator/pkg/cps/init.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"context"
 
-	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-operator/pkg/apis/galasa/v1alpha1"
+	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-kubernetes-operator/pkg/apis/galasa/v1alpha1"
 
 	corev1 "k8s.io/api/core/v1"
