+++ galasa-ecosystem-kubernetes-operator/pkg/ras/ras_test.go
@@ -4,7 +4,7 @@ import (
 	"strconv"
 	"testing"
 
-	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-operator/pkg/apis/galasa/v1alpha1"
+	galasav1alpha1 "github.com/galasa-dev/extensions/galasa-ecosystem-kubernetes-operator/pkg/apis/galasa/v1alpha1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
